requirements.txt
accelerate==0.31.0
diffusers==0.29.2
gradio==4.39.0
spaces==0.29.2
huggingface_hub==0.23.4
matplotlib==3.9.1
numpy==1.26.4
opencv_python==4.10.0.84
pillow==10.3.0
PyYAML==6.0.1
scipy==1.13.1
setuptools==51.0.0
scikit-image==0.24.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.4
transformers==4.27.3
git+https://github.com/facebookresearch/detectron2.git
git+https://github.com/facebookresearch/detectron2@main#subdirectory=projects/DensePose
Ninja==1.11.1